fix: bug in number conversion

+++ src/geometry.cpp
@@ -1,7 +1,7 @@
 #include "geometry.h"
 
 Vec2D::Vec2D() : x(0.0f), y(0.0f) {}
-Vec2D::Vec2D(float x, float y) : x(x), y(y) {}
+Vec2D::Vec2D(int x, int y) : x(x), y(y) {}
 Vec2D Vec2D::operator+(Vec2D const& rhs) { return Vec2D(x + rhs.x, y + rhs.y); }
 Vec2D& Vec2D::operator+=(Vec2D const& rhs) {
   x += rhs.x;
@@ -9,6 +9,6 @@ Vec2D& Vec2D::operator+=(Vec2D const& rhs) {
 
   return *this;
 }
-Vec2D Vec2D::operator*(float rhs) { return Vec2D(x * rhs, y * rhs); }
+Vec2D Vec2D::operator*(int rhs) { return Vec2D(x * rhs, y * rhs); }
 
 Vec2D::~Vec2D() {}
