+++ src/geometry.h
@@ -10,11 +10,11 @@ struct Geometry {
 class Vec2D {
  public:
   Vec2D();
-  Vec2D(float x, float y);
+  Vec2D(int x, int y);
   ~Vec2D();
   Vec2D operator+(Vec2D const& rhs);
   Vec2D& operator+=(Vec2D const& rhs);
-  Vec2D operator*(float rhs);
+  Vec2D operator*(int rhs);
 
   int x{0}, y{0};
 };
